src/main/java/com/kimyounghan/hellospring/repository/MemoryMemberRepository.java
package com.kimyounghan.hellospring.repository;

import com.kimyounghan.hellospring.domain.Member;
import org.springframework.stereotype.Repository;

import java.util.*;

@Repository
public class MemoryMemberRepository implements MemberRepository{

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    /**
     *    매개변수로 전달된 Member 객체의 id값을 지정해주고 store에 저장하는 메서드
     *    @Param Member
     */
    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {

        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }
    /**
     *    store에 저장되어있는 hashmap 요소를 List 객체로 변환하여 반환하는 메서드
     *    @Return List
     */
    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }
}
